refactor: small refactoring of responsibilities
